% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwcomparator.R
\name{aligninputandrefdata}
\alias{aligninputandrefdata}
\title{This function translate inputdata so that it has common values of limbname Dx, Dy, Dz columns with refdata.}
\usage{
aligninputandrefdata(inputdata, refdata, limbname)
}
\arguments{
\item{inputdata}{input motion capture data frame.}

\item{refdata}{reference motion capture data frame.}

\item{limbname}{name of the column which is used to perform alignment.}
}
\value{
data frame aligned as it is described above.
}
\description{
This function translate inputdata so that it has common values of limbname Dx, Dy, Dz columns with refdata.
}
\examples{
#See example from analyze.mocap function
}
