% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dba_quaternion_averaging_speed_up_all_techniques.R
\docType{class}
\name{averaged.mocap-class}
\alias{averaged.mocap-class}
\alias{averaged.mocap}
\title{A class returned by mocap.averaging function.}
\format{a list containing data frame (fullData) and data frame (norm.distance) with normalized distance optimized during averaging.}
\usage{
see documentation of mocap.averaging.
}
\description{
A class returned by mocap.averaging function.
}
\examples{
data("mawashi.geri.right.list")
myList <- list()
for (a in 1:length(mawashi.geri.right.list))
{
  myList[[a]] <-mawashi.geri.right.list[[a]]$data.frame
}
res.data <- mocap.averagingCmp(myList, 2, eps = 0.000001)
plot(res.data)
}
\keyword{class}
