% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{bvh.to.csv}
\alias{bvh.to.csv}
\title{This function get direct kinematic from list generated with function read.bvh or read.mocap function and returns it in the form of data frame.}
\usage{
bvh.to.csv(skeleton, sd = TRUE)
}
\arguments{
\item{skeleton}{input hierarchical kinematic model.}

\item{sd}{does function should calculate second derivative? Default = TRUE.}
}
\value{
data frame with direct kinematic.
}
\description{
This function does not perform any algebraic calculation - it just takes data from Dxyz and Rxyz columns. Function can additionally calculate second derivative of Dxyz.
}
\examples{
#an example BVH file
data("heian.nidan.bvh")
#write file to the disc
f <- file("heian.nidan.bvh")
writeChar(con = f, object = heian.nidan.bvh)
close(f)
#read hierarchical model stored in BVH file
}
