% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{calculate.kinematic}
\alias{calculate.kinematic}
\title{This function corrects translation of the direct kinematic model.}
\usage{
calculate.kinematic(dd, LeftFoot = "LeftFoot", RightFoot = "RightFoot",
  show.plot = FALSE, plot.title = "", bodypartname = NULL, dyEps = 5)
}
\arguments{
\item{dd}{data frame with motion capture data.}

\item{LeftFoot}{name of the column with data that holds coordinates of left foot (joint that touch the ground). Default value is LeftFoot = "LeftFoot". Used only for the first type of call.}

\item{RightFoot}{name of the column with data that holds coordinates of right foot (joint that touch the ground). Default value is RightFoot = "RightFoot". Used only for the first type of call.}

\item{show.plot}{if TRUE (default is show.plot = FALSE) plots results of an algorithm.}

\item{plot.title}{part of the title over plots.}

\item{bodypartname}{if not NULL value the stationary joint of the motion is column with bodypartname. Default value is bodypartname = NULL.}

\item{dyEps}{Treshold value for translation calculation (used only for the first type of call). Default value is dyEps = 5.}
}
\value{
Data frame, which has Dxzy columns updated. All other columns are not updated.
}
\description{
This heuristic method is especially usable while dealing with data acquired by IMU (Inertial measurement unit) sensors. There are two
possible calls of this method. The first one utilize acceleration data of body joints (.ax, .ay and .az columns). If this data is avilable
method can calculate displacement of long motion, during which an actor uses both left and right leg. If this call is used,
one should left bodypartname as NULL. The second call is used when we are dealing with data without acceleration data and actor has
one stationary limb. In this case bodypartname shoule have value of the limb which does not translate during motion.
}
\examples{
  #This example uses acceleration data to calculates correct displacement.
  #header of mocap file
  data("header.mocap")
  #data frame with displacements and acceleration data
  data("heian.shodan")

  heian.shodan.corrected <- calculate.kinematic(heian.shodan, show.plot = "TRUE", plot.title = "Heian Shodan")
  original.bvh <- set.data.frame(header.mocap, heian.shodan)
  corrected.bvh <- set.data.frame(header.mocap, heian.shodan.corrected)
  #plot BVH, red is original data, green is corrected
  plot(original.bvh, frames.fraction = 0.1, my.color = "red", alpha = 0.1, spheres = FALSE)
  plot(corrected.bvh, frames.fraction = 0.1, my.color = "green", alpha = 0.1, spheres = FALSE, append = TRUE)

  #writting BVH to disk
  write.bvh(original.bvh, "original.bvh")
  write.bvh(corrected.bvh, "corrected.bvh")

  #This example uses single body joint which coordinates should be constant during motion
  data(mawashi.geri.left.1)
  plot(mawashi.geri.left.1, frames.fraction = 0.1, my.color = "red", alpha = 0.5, spheres = FALSE)
  mawashi.geri.left.1$data.frame <- calculate.kinematic(mawashi.geri.left.1$data.frame, bodypartname = "RightFoot")
  plot(mawashi.geri.left.1, frames.fraction = 0.1, my.color = "green", alpha = 0.5, spheres = FALSE, append = TRUE)
}
