% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwcomparator.R
\name{euc.dist}
\alias{euc.dist}
\title{This function calculates Euclidean distance between vectors x1 and x2.}
\usage{
euc.dist(x1, x2)
}
\arguments{
\item{x1}{first numeric vector.}

\item{x2}{second numeric vector.}
}
\value{
asqrt(sum((x1 - x2) ^ 2)).
}
\description{
This function calculates Euclidean distance between vectors x1 and x2.
}
\examples{
euc.dist(c(1,2,3,4),c(-5,0,-6, 3))
}
