% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwcomparator.R
\name{euc.dist1d}
\alias{euc.dist1d}
\title{This function returns absolute value of difference between x1 and x2.}
\usage{
euc.dist1d(x1, x2)
}
\arguments{
\item{x1}{first numeric value.}

\item{x2}{second numeric value.}
}
\value{
abs(x1[1] - x2[1]).
}
\description{
This function returns absolute value of difference between x1 and x2.
}
\examples{
euc.dist1d(1,-5)
}
