% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{generate.first.frame}
\alias{generate.first.frame}
\title{This function generates object of mocap class with zero rotation data.}
\usage{
generate.first.frame(input.skeleton, Nframes = 1, FrameTime = 0.01)
}
\arguments{
\item{input.skeleton}{object of mocap class which defines hierarchical model.}

\item{Nframes}{number of frames to be generated.}

\item{FrameTime}{intervals between acquisitions.}
}
\value{
object of mocap class.
}
\description{
This function generates object of mocap class with zero rotation data.
}
\examples{
data("heian.nidan.bvh")
f <- file("heian.nidan.bvh")
writeChar(con = f, object = heian.nidan.bvh)
close(f)
#read hierarchical model stored in hierarchical BVH file
heian.nidan <- read.mocap("heian.nidan.bvh")
first.frame <- generate.first.frame(heian.nidan)
plot(heian.nidan, frame = 1, alpha = 1, spheres = TRUE)
plot(first.frame, my.color = "red", frame = 1, alpha = 1, spheres = TRUE, append = TRUE)
}
