% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\docType{data}
\name{header.mocap}
\alias{header.mocap}
\title{An example object of class mocap that does not have any motion data.}
\format{An object of mocap class.}
\usage{
data("header.mocap")
}
\description{
An example object of class mocap that does not have any motion data.
}
\keyword{datasets}
