% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\docType{data}
\name{heian.nidan.bvh}
\alias{heian.nidan.bvh}
\title{A raw bvh file to be saved on disc (Shotokan Karate kata Heian Nidan).}
\format{A raw file.}
\usage{
data(heian.nidan.bvh)
}
\description{
A raw bvh file to be saved on disc (Shotokan Karate kata Heian Nidan).
}
\examples{
data("heian.nidan.bvh")
f <- file("e:\\\\bvh in r\\\\gotowy_kod\\\\output\\\\heian.nidan.bvh")
writeChar(con = f, object = heian.nidan.bvh)
close(f)
}
\keyword{datasets}
