% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\docType{data}
\name{heian.shodan}
\alias{heian.shodan}
\title{A data frame with mocap data (Shotokan Karate kata Heian Shodan). Also acceleration channels are included.}
\format{A data frame.}
\usage{
data("heian.shodan")
}
\description{
A data frame with mocap data (Shotokan Karate kata Heian Shodan). Also acceleration channels are included.
}
\keyword{datasets}
