% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\docType{data}
\name{heian.yondan}
\alias{heian.yondan}
\title{A data frame with mocap data (Shotokan Karate kata Heian Yondan).}
\format{A data frame.}
\usage{
data("heian.yondan")
}
\description{
A data frame with mocap data (Shotokan Karate kata Heian Yondan).
}
\keyword{datasets}
