% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{hierarchical.to.direct.kinematic}
\alias{hierarchical.to.direct.kinematic}
\title{Generates direct kinematic model from hierarchical kinematic model.}
\usage{
hierarchical.to.direct.kinematic(input.skeleton)
}
\arguments{
\item{input.skeleton}{list in the same format as one generated with read.mocap function.}
}
\value{
data frame with direct kinematic model. Names of the columns are the same as in input.skeleton.
}
\description{
This function makes calculations based on hierarchical kinematic data in input list (input.skeleton). It does not use Dxyz from input.skeleton.
}
\examples{
#an example BVH file
data("heian.nidan.bvh")
f <- file("heian.nidan.bvh")
writeChar(con = f, object = heian.nidan.bvh)
close(f)
#read hierarchical model stored in hierarchical BVH file
heian.nidan <- read.mocap("heian.nidan.bvh")
#plot kinematic data
plot(x = heian.nidan$data.frame$Hips.Dx, y = heian.nidan$data.frame$Hips.Dz, type = "l", ylab = "Displacement X [cm]", xlab = "Displacement Z [cm]")
title("Hips displacement during motion")

#generate kinematic from hierarchical model - same results as above
df <- hierarchical.to.direct.kinematic(heian.nidan$skeleton)
plot(x = df$Hips.Dx, y = df$Hips.Dz, type = "l", ylab = "Displacement X [cm]", xlab = "Displacement Z [cm]")
title("Hips displacement during motion")
}
