% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\docType{data}
\name{mawashi.geri.left.2}
\alias{mawashi.geri.left.2}
\title{An object of class mocap that contains karate kick mawashi geri.}
\format{An object of mocap class.}
\usage{
data("mawashi.geri.left.2")
}
\description{
An object of class mocap that contains karate kick mawashi geri.
}
\keyword{datasets}
