% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\docType{data}
\name{mawashi.geri.right.list}
\alias{mawashi.geri.right.list}
\title{A list of ten object of class mocap that contains karate kicks mawashi geri.}
\format{A list of objects of mocap class.}
\usage{
data("mawashi.geri.right.list")
}
\description{
A list of ten object of class mocap that contains karate kicks mawashi geri.
}
\keyword{datasets}
