% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dba_quaternion_averaging_speed_up_all_techniques.R
\name{plot.averaged.mocap}
\alias{plot.averaged.mocap}
\title{Plots normalized distance of all joints from averaged.mocap class.}
\usage{
\method{plot}{averaged.mocap}(data.to.plot)
}
\arguments{
\item{data.to.plot}{object of class averaged.mocap.}
}
\description{
Plots normalized distance of all joints from averaged.mocap class.
}
\examples{
data("mawashi.geri.right.list")
myList <- list()
for (a in 1:length(mawashi.geri.right.list))
{
  myList[[a]] <-mawashi.geri.right.list[[a]]$data.frame
}
res.data <- mocap.averagingCmp(myList, 2, eps = 0.000001)
plot(res.data)
}
