% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{plot.mocap}
\alias{plot.mocap}
\title{Overrides plot function to work with mocap class.}
\usage{
\method{plot}{mocap}(obj, frame = 0, my.color = "green",
  frames.fraction = 0.1, alpha = 0.05, spheres = FALSE, append = FALSE,
  print.text = FALSE)
}
\arguments{
\item{obj}{mocap object.}

\item{frame}{index of frame to be show. If default (frame = 0) all frames are drawn.}

\item{my.color}{color of the plot.}

\item{frames.fraction}{if frame = 0 this parameter indicates what fraction of frames should be drawn (default frames.fraction = 0.1 means that 10\% of frames will be plotted).}

\item{alpha}{value of alpha channel of the plot (0 is 100\% transparency, 1 is no transparency, default is alpha = 0.05).}

\item{spheres}{if TRUE, position of body joints will be marked as spheres (default is spheres = FALSE).}

\item{append}{if FALSE (default is append = FALSE) a new plot is generated, if TRUE a plot is drawn over open rgl window).}

\item{print.text}{if TRUE (default is append = FALSE) plots name of the joints.}
}
\description{
This function uses rgl package for 3D visualizations. It creates interactive 3D plots that enables rotation and scaling.
}
\examples{
data("right.arm.motion.1")
plot(right.arm.motion.1, frame = 1, my.color = "white", alpha = 1, spheres = TRUE)
plot(right.arm.motion.1, frames.fraction = 0.5, my.color = "white", alpha = 1, spheres = FALSE, print.text = TRUE)
}
