% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{read.bvh}
\alias{read.bvh}
\title{This function read file in BVH (Biovision Hierarchy) format and recalculates it to direct kinematic model.}
\usage{
read.bvh(filepath)
}
\arguments{
\item{filepath}{A path to BVH file.}
}
\value{
an object of mocap class.
}
\description{
BVH file contains a hierarchical kinematic model definition (called a skeleton) and motion data. For more information see:
"Motion Capture File Formats Explained" by M. Meredith S. Maddock, doi: 10.1.1.103.2097.
}
\examples{
#an example BVH file
data("heian.nidan.bvh")
#write file to the disc
f <- file("heian.nidan.bvh")
writeChar(con = f, object = heian.nidan.bvh)
close(f)
#read hierarchical model stored in BVH file
}
