% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{read.mocap}
\alias{read.mocap}
\title{This function reads motion capture file on BVH format.}
\usage{
read.mocap(filepath)
}
\arguments{
\item{filepath}{path to a file.}
}
\value{
object of mocap class.
}
\description{
It also calculates direct kinematic from hierarchical kinematic. This function calls read.bvh and bvh.to.df to generate single object of mocap class. See documentation for those two functions.
}
\examples{
#an example BVH file
data("heian.nidan.bvh")
f <- file("eian.nidan.bvh")
writeChar(con = f, object = heian.nidan.bvh)
close(f)
#read hierarchical model stored in hierarchical BVH file
heian.nidan <- read.mocap("heian.nidan.bvh")
summary(heian.nidan)
}
