% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\docType{data}
\name{right.arm.motion.2}
\alias{right.arm.motion.2}
\title{An object of class mocap that contains motion of right arm.}
\format{An object of mocap class.}
\usage{
data("right.arm.motion.2")
}
\description{
An object of class mocap that contains motion of right arm.
}
\keyword{datasets}
