% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{rotation.matrix.between.vectors}
\alias{rotation.matrix.between.vectors}
\title{This function returns n by n rotation matrix that align vector x onto y.}
\usage{
rotation.matrix.between.vectors(x, y)
}
\arguments{
\item{x}{vector to be rotated (has to be normalized).}

\item{y}{reference vector (has to be normalized).}
}
\value{
n x n rotation matrix.
}
\description{
Both vector x and y has to be normalized.
}
\examples{
x <- vector.to.unit(c(1,0,0))
y <- vector.to.unit(c(0,1,0))
Rx2y <- rotation.matrix.between.vectors(x, y)
x \%*\% Rx2y
y
x <- vector.to.unit(c(5,0,2,4,6))
y <- vector.to.unit(c(0,1,0,8,-5))
Rx2y <- rotation.matrix.between.vectors(x, y)
x \%*\% Rx2y
y
x <- vector.to.unit(c(8,0,0))
y <- vector.to.unit(c(2,0,0))
Rx2y <- rotation.matrix.between.vectors(x, y)
x \%*\% Rx2y
y
}
