% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{second.derivative}
\alias{second.derivative}
\title{Calculates second derivative}
\usage{
second.derivative(signal)
}
\arguments{
\item{signal}{input vecotr.}
}
\value{
vector containing second derivative calculated from the input vector.
}
\description{
Calculates second derivative using Central Difference Operator.
}
\examples{
second.derivative(sin(seq(from=0,to=2*pi,by=pi/50)))
}
