% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{summary.mocap}
\alias{summary.mocap}
\title{Plots information about number of frames, joints count and joints chierarchy of mocap class.}
\usage{
\method{summary}{mocap}(mocap.data)
}
\arguments{
\item{mocap.data}{an object of mocap class.}
}
\description{
Plots information about number of frames, joints count and joints chierarchy of mocap class.
}
\examples{
data("heian.nidan.bvh")
f <- file("e:\\\\bvh in r\\\\gotowy_kod\\\\output\\\\heian.nidan.bvh")
writeChar(con = f, object = heian.nidan.bvh)
close(f)
#read hierarchical model stored in hierarchical BVH file
heian.nidan <- read.mocap("e:\\\\bvh in r\\\\gotowy_kod\\\\output\\\\heian.nidan.bvh")
summary(heian.nidan)
}
