% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{vector.angle}
\alias{vector.angle}
\title{This function calculates angle between two vectors on the plane designed by those vectors.}
\usage{
vector.angle(a, b)
}
\arguments{
\item{a}{first vector.}

\item{b}{second vector.}
}
\value{
angle between a and b (in radians).
}
\description{
The return value is in the range [0,pi].
}
\examples{
vector.angle(c(1,2,3),c(4,5,6))
}
