% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{vector.cross}
\alias{vector.cross}
\title{This function calculates cross product.}
\usage{
vector.cross(a, b)
}
\arguments{
\item{a}{first vector.}

\item{b}{second vector.}
}
\value{
cross product.
}
\description{
Cross product is only defined for 3D vectors.
}
\examples{
vector.cross(c(1,2,3),c(3,4,5))
}
