% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{vector.dot}
\alias{vector.dot}
\title{This function calculates dot product.}
\usage{
vector.dot(a, b)
}
\arguments{
\item{a}{first vector.}

\item{b}{second vector.}
}
\value{
dot product.
}
\description{
This function calculates dot product.
}
\examples{
vector.dot(c(1,2,3,4),c(5,6,7,8))
}
