% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{vector.norm}
\alias{vector.norm}
\title{This function returns length (norm) of the vector.}
\usage{
vector.norm(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
vector norm.
}
\description{
This function returns length (norm) of the vector.
}
\examples{
vector.norm(c(1,2,3))
}
