% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{vector.to.angles.frame.list}
\alias{vector.to.angles.frame.list}
\title{This function generates list of angles between vector and coordinates frame.}
\usage{
vector.to.angles.frame.list(dataToCalculate, f1, f2, xt1, xt2)
}
\arguments{
\item{dataToCalculate}{data frame with motion capture data.}

\item{f1}{name of the first joint of vector.}

\item{f2}{name of the second joint of vector.}

\item{xt1}{name of the first joint of horizontal axis of coordinate frame.}

\item{xt2}{name of the second joint of horizontal axis of coordinate frame.}
}
\value{
Results are list of vectors defined as follow: (angle(Y, v1), angle(Z, v1), angle(X, v1)).
}
\description{
Vector is defined as v1=(f1-f2). The coordinate frame is:
\itemize{
  \item X=(xt1-xt2),
  \item Z=X x (0,1,0),
  \item Y=X x Z
}
All vectors are normalized.
}
\examples{
data("heian.yondan")
vector.to.angles.frame.list(heian.yondan[1:3,], "RightArm", "RightForearm", "RightShoulder", "LeftShoulder")
}
