% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{vector.to.angles.list}
\alias{vector.to.angles.list}
\title{This function generates list of angles.}
\usage{
vector.to.angles.list(dataToCalculate, f1, f2, f3)
}
\arguments{
\item{dataToCalculate}{data frame with motion capture data.}

\item{f1}{name of the first joint.}

\item{f2}{name of the second joint.}

\item{f3}{name of the third joint.}
}
\value{
list of angles (in radians) defined as above.
}
\description{
Angles are calculated on the plane between vectors v1=(f2-f1) and v2=(f2-f3).
}
\examples{
data("heian.yondan")
vector.to.angles.list(heian.yondan[1:3,], "RightShoulder", "RightArm", "RightForearm")
}
