% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{vector.to.list}
\alias{vector.to.list}
\title{This function genereates list of vectors.}
\usage{
vector.to.list(dataToCalculate, featuresName)
}
\arguments{
\item{dataToCalculate}{data frame with motion capture data.}

\item{featuresName}{column name of the motion capture feature. Column names should have names endings .Dx, .Dy, .Dz.}
}
\value{
list of vectors defined as above.
}
\description{
Each list position equals c(ABC.Dx, ABC.Dy, ABC.Dz) where featuresName = ABC is a name of the column of data frame dataToCalculate.
}
\examples{
data("heian.yondan")
vector.to.list(heian.yondan[1:3,], "RightHand")
}
