% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{vector.to.unit}
\alias{vector.to.unit}
\title{This function performs vector normalizing.}
\usage{
vector.to.unit(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
normalized vector x.
}
\description{
This function performs vector normalizing.
}
\examples{
vector.to.unit(c(1,2,3,4,5))
}
