% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvhutils.R
\name{write.bvh}
\alias{write.bvh}
\title{This function saves object of mocap class to a file.}
\usage{
write.bvh(skeleton.helper, path)
}
\arguments{
\item{skeleton.helper}{mocap object to be save.}

\item{path}{path to the file.}
}
\description{
Function uses hierarchical model definition from hierarchy list however data of channels are taken from data frame.
}
\examples{
  data("header.mocap")
  data("heian.shodan")
  heian.shodan.corrected <- calculate.kinematic(heian.shodan, show.plot = "TRUE", plot.title = "Heian Shodan")
  original.bvh <- set.data.frame(header.mocap, heian.shodan)
  corrected.bvh <- set.data.frame(header.mocap, heian.shodan.corrected)
  #plotting BVH
  plot(original.bvh, frames.fraction = 0.1, my.color = "red", alpha = 0.1, spheres = FALSE)
  plot(corrected.bvh, frames.fraction = 0.1, my.color = "green", alpha = 0.1, spheres = FALSE, append = TRUE)
  #writing BVH to disk
  write.bvh(original.bvh, "original.bvh")
  write.bvh(corrected.bvh, "corrected.bvh")
}
